requirements.txt
requests~=2.31.0
streamlit~=1.31.1
streamlit_extras~=0.4.2
streamlit_chatbox~=1.1.12
validators~=0.22.0
python-dotenv~=1.0.1
pillow~=10.3.0
pyyaml~=6.0.1
openai~=1.30.1
typing_extensions~=4.7.1
simplejson~=3.19.2

langchain~=0.2.0
langchain-core~=0.2.0
langchain-openai~=0.1.7
langchain-wenxin~=0.10.2

lark_oapi~=1.2.5
oss2~=2.18.5